+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add support for arrays of `html_safe` strings
+
 ## [0.1.0] - 2023-01-08
 
 - Initial release